+++ KoS.Apps.SharePoint.SmartCAML/changelog.md
@@ -1,6 +1,10 @@
 ## Changelog
 #### v. 1.8.0
-+ New application icon and logo
+##### New
++ Application icon and logo
 + Send update and install information to the server
 + Support 'Membership' operator
 + Support 'In' Operator
+##### Fix
++ Display DateTime values in local time zone
++ Updating DateTime field now will no longer crach